Reverted a585e1bd... requirejs fails to build the latest ember(-data) versions
The internal module loader structure of ember and ember data has changed and causes requirejs to fail. This can be fixed by replacing both module names or by replacing the name of the internal module loaders. This is quite ugly and I rather wait for a fix.

Also changed jquery version to address the old dependency issue...