Created initializers with dependency injections
The application metadata and settings records were moved to their own
initializers and will now be injected into all routes and controllers
rather than being set during the beforeModel, model and setupController
hooks of the ApplicationRoute.
Dependency injections are still WIP...

Thanks to the DI, the logic for the initial redirection to the custom
homepage could be moved to the IndexRoute.