Fix modal popup animation
Sometimes the modal element gets inserted into the DOM with the `shown`
class added, in which case the animation doesn't play. The fix is adding
a slight delay of 20ms to the class toggle after the next Ember run loop
iteration, so that we can be sure that the element was inserted.