Implement custom livestreamer qualities. #323
Add all currently available explicit stream qualities to the value of
the "source" quality in descending order. This ensures that streams
can be opened, even if "source" is missing from their quality list.

This commit changes the type of the quality attribute of the
Settings and ChannelSettings models from number to string.