Remove file ext from the which function on Windows
This change requires all Windows users to set filenames explicitly.
Since the fileselect component is setting the whole file name and since
most users probably won't know that the .exe extension is optional, this
won't be a big deal.

This reduces the promise chain by a bit and enables checking for files
in $PATH without the .exe file name extension.