Fix broadcaster_language
Since 63b38ad, ObjectBuffer instances keep references of their
nested objects when applying changes and reading their content.

Add a target param to ObjectBuffer.applyChanges that gets updated with
the new content and gets notified when nested objects were modified.

This fixes the observer of the LanguageFilterMixin, which was not being
notified when properties of the gui_langfilter objects got changed.