Refactor streaming provider validation
Streamlink portable could be added as a separate streaming provider
as an addition to #451 (which was the motivation to this commit), but
a Windows command prompt window pops up when the executable gets
launched from the GUI, which makes it unusable. Stdout and stderr are
not being passed through and can't be read.