Selective dependency resolution for just-extend
Fixes security alert and keeps nise on version 1.3.3, so that the usage
of sinon in tests doesn't trigger unnecessary deprecation warnings of
Ember's various prototype extensions, which are nowhere used else.

The warnings are triggered by sinon due to its stubbing logic and the
iteration of the object prototype properties. Since Ember didn't declare
the deprecated properties as enumerable: false, sinon and nise can not
be upgraded if we don't want unnecessary deprecation messages in tests.