Disable nwjs zip option and remove platformfixes
This disables the zip option in all builds (affected Windows and Linux),
so that the application code doesn't get zipped and appended to the
NW.js executable which it then needs to extract again on each run into
a temp directory. Since the app code lived in a temp dir on these platforms,
the CWD of the process had to be changed in the platformfixes module and
every relative path resolution had to be worked around by adding a temporary
NWJSAPPPATH env var.