Simplify streaming providers
Resolves #718

- rename "streamlink" to "streamlink-python"
- rename "streamlinkw" to "streamlink" and add data for macOS and Linux
- add localstorage transform logic (keep settings)
- fix default-provider not being read from streaming config

This makes the "streamlink" resolver config the default and simply
launches the "streamlink" or "streamlinkw.exe" executable instead of
trying to resolve the python entry module file and its associated
python executable.

This python resolver was needed in the past due to Streamlink having a
hardcoded executable on Windows which always spawned a terminal window.
The workaround was reading the python path from the shebang of the entry
script and replacing the python executable name and executing this
custom path instead with the entry script as first parameter.

However, this resolving workaround is not needed anymore due to a patch
in pynsist and a config change in Streamlink's wheel distribution, which
enabled Streamlink having two separate executables. Since the minimum
version requirement of Streamlink is already 1.4.0, backwards
compatibility does not need to be kept.

This change should be transparent on all OSes, except in those cases
where custom python paths and entry scripts were !intentionally! set.
Here, the streaming provider needs to be changed to "streamlink-python"
to use the old settings.