build: add debug build configs
This adds the `build:debug` task and the `nwjs`, `release`, `compile`
and `dist` tasks now support the trailing `debug` task argument, which
will compile the debug build instead of the production build. Example:
  grunt build:debug dist:linux64:debug

This also adds the new `DEVELOPMENT` global var in addition to `DEBUG`
and splits some of the logic between those two, as some code is meant
to only be run in a development build.

Debug builds have the regular debug flags sets, similar to the
development build, and the NW.js SDK flavor is used for compiling them,
so that the Chromium dev-tools can be accessed.