services/streaming: clean up on modal quit dialog
This automatically closes the ended stream's modal if the quit modal
dialog is shown, so that the stream gets removed from the store.
This will make the quit modal dialog update correctly if a stream gets
closed while the modal is shown, even though the stream's modal was
still opened in the background.