init/instance: fix NW.js window initialization
Resolves #911

Don't call `requestAnimationFrame` while the application window is still
hidden during initialization. NW.js sometimes doesn't execute the
animation frame callback during the initial window visibility state.
This became clear after the NW.js 0.68.1 upgrade recently, which lead to
the initialization never completing in certain cases on some systems.

Remove the two `requestAnimationFrame` calls and schedule the
`afterRender` Ember run-loop callback in the next run-loop, which
ensures that the DOM is actually fully loaded and rendered by NW.js, so
that no white screen appears for a few frames, which needs to be avoided
when using the dark theme.