ui/components/modal-streaming: fix close on error
Fixes #954

Destroy the `Stream` record when closing the streaming modal if the
stream has already ended, don't just try to kill the process. Use the
same `streaming.closeStream()` call which the `WatchingController` does.

The streaming modal gets closed automatically if the stream has ended
regularly (unless disabled in the settings), but it's still possible
to open the modal dialog from the watching menu later on when it ended
in an error (e.g. read timeout).