services/streaming: conditionally restore GUI
Resolves #955

Refactor `StreamingService`, add a separate GUI settings option for
restoring the GUI window after one or all streams have ended, and make
the minimize and restore settings/functions independent from each other.

Set the default value of the new settings option to "no-op" and don't
restore the hidden GUI window when streams end. This differs from
the previous implementation where the GUI window would be restored
if any of the minimize or move-to-tray options was selected.