build: fix assets in parse-json-loader
The switch towards asset modules in 242472b broke the parse-json-loader,
leading to missing icon files when reading the files.json config.

This is a known bug in webpack 5, which doesn't emit asset files when
loading asset modules via `LoaderContext.loadModule()`. Unfortunately,
the webpack developers decided against fixing this last year after
opening a bug report with a reproduction and explaining the issue.

Fix it by explicitly calling `this.emitFile()` in the `loadModule()`
callback and fix the variable scope when setting the resolved path.