vendor: bump babel/traverse
and finally resolve dependabot alert

Don't bump `babel/core` and don't deduplicate `babel/generator`,
as it causes the `test:coverage` run to never finish on GH's CI runners.

The reason for this is likely some kind of incompatibility with
`babel-plugin-instanbul` and its code instrumentation. This still
needs to be debugged and solved in the future.

Alternatively, istanbul-js could be replaced with Chromium's
native v8 code coverage feature and reading the data from the CDP.
See #988