nwjs/tray: fix attaching menu to new tray icon
When "closing to tray" and re-launching the application, the sent
"open" event always rebuilds the tray icon when "external commands"
and advanced settings are enabled. The existing tray icon menu object
is re-used in this case and gets attached to the new tray icon object.

For some reason though, the menu is sometimes missing, resulting in
a state where the user can't close the application anymore without
changing the tray icon settings. This is apparently caused by a timing
issue in NW.js.

Increasing the delay when re-attaching the menu seems to fix the issue.