build: add Linux wrapper launch script
Add wrapper launch script in place of the old executable name
and set `--ozone-platform=wayland` if `XDG_SESSION_TYPE` is `wayland`.

This prevents NW.js to be run in an XWayland session, which is currently
still the default behavior of Chromium.

TODO: remove this wrapper launch script once changed in Chromium